Several compiler warning fixes:
     * msp430: fix "implicit declaration of function" warnings in clock.c, by including watchdog.h;
     * sky: fix a couple pointer target signedness warnings;
     * core: fix several signed/unsigned comparison warnings;
     * framer-802154: "const static" -> "static const" to fix compiler warnings;
     * core: comment or remove unused variables and function definitions.

+++ core/net/mac/cxmac.c
@@ -454,7 +454,7 @@ send_packet(void)
     packetbuf_attr(PACKETBUF_ATTR_ERELIABLE);
   len = NETSTACK_FRAMER.create();
   strobe_len = len + sizeof(struct cxmac_hdr);
-  if(len == 0 || strobe_len > sizeof(strobe)) {
+  if(len == 0 || strobe_len > (int)sizeof(strobe)) {
     /* Failed to send */
    PRINTF("cxmac: send failed, too large header\n");
     return MAC_TX_ERR_FATAL;
